requirements.txt
streamlit
google-genai
pypdf
langchain-core
langchain-community
langchain-google-genai